refactor(anyharness): group live terminal modules by role
