+++ anyharness/crates/anyharness-lib/src/live/terminals/command_runs/mod.rs
@@ -1,0 +1,8 @@
+mod pty;
+mod setup_process;
+mod stream_format;
+
+pub(super) use pty::{process_pty_output, run_terminal_command, ActivePtyCommand};
+pub(super) use setup_process::{
+    run_setup_process, set_terminal_output_suppressed, ActiveSetupTask,
+};
