+++ anyharness/crates/anyharness-lib/src/live/terminals/command_runs/stream_format.rs
