+++ anyharness/crates/anyharness-lib/src/live/terminals/driver/mod.rs
@@ -1,0 +1,5 @@
+mod pty;
+mod shell;
+
+pub(super) use pty::create_terminal_shell;
+pub(super) use shell::detect_posix_shell;
