+++ anyharness/crates/anyharness-lib/src/live/terminals/mod.rs
@@ -1,13 +1,9 @@
 mod agent_login;
+mod command_runs;
 mod driver;
 mod handle;
 mod manager;
 mod output_sink;
-mod pty_command;
-mod replay;
-mod setup_process;
-mod shell;
-mod stream_format;
 
 pub use crate::domains::terminals::model::TerminalOutputEvent;
 pub use agent_login::{
